Corrected file name and imports for sensor_transformation_utils.py
The file was erroneously called sensor_transformation_utils.py.py (with the .py extension repeated), so I fixed it to have only one ".py".

Also, the script attempts to import the Transform module from opencood.utils.opencda_carla, which does not exist in this repository. Note that this script actually defines its own Transform class later on, so the import statement on Line 12 seems unnecessary, and breaks the script.